feat: Add new modules for GKE and GLB
Add new files and configurations for the GKE and GLB modules. This includes the creation of the GKE cluster and node pool, as well as the configuration of the GLB with backend services and forwarding rules.

feat: Add new VPC module

Add a new VPC module with the necessary resources for creating a VPC network, subnet, and secondary IP ranges. This module will be used by the GKE module.

feat: Add deployment and service configurations

Add deployment and service configurations for the Kubernetes application. This includes the deployment of a demo app and the creation of a service with a ClusterIP type.

feat: Add Terraform backend configuration

Add the necessary configuration for using Google Cloud Storage as the Terraform backend. This includes specifying the bucket and prefix for storing the Terraform state.

feat: Add GitHub Actions workflow

Add a GitHub Actions workflow for automating the deployment of the infrastructure. This workflow will trigger on pushes