feat: Add disk_size_gb variable to GKE module
Add a new variable "disk_size_gb" to the GKE module in the file deployments/terraform/modules/gke/variables.tf. This variable represents the size of the disk attached to each node in the cluster's default node pool. The variable is of type number and has a default value of 10.

Also, update the GKE module in the file deployments/terraform/modules/gke/main.tf to include the new "disk_size_gb" variable in the node_config block. The value of "disk_size_gb" is set to the value of the variable passed in.

This change adds flexibility to the GKE module by allowing users to customize the size of the disk attached to each node in the cluster.